RealEstate/requirements.txt
cycler==0.10.0
decorator==4.4.2
Django==3.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4

Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.1.2

psycopg2 == 2.8.6

python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5

sqlparse==0.4.1

whitenoise==5.2.0

waitress==2.0.0

